src/auth/requirements.txt
dependency-injector==4.41.0
fastapi==0.103.2
fastapi-versioning==0.10.0
flake8==7.0.0
gunicorn==20.1.0
httpx==0.26.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
PyYAML==6.0.1
motor==3.3.2
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.22.0
